+++ web/modules/custom/ai_screening_project_track/src/ProjectTrackToolAccessControlHandler.php
@@ -6,8 +6,12 @@ namespace Drupal\ai_screening_project_track;
 
 use Drupal\Core\Access\AccessResult;
 use Drupal\Core\Entity\EntityAccessControlHandler;
+use Drupal\Core\Entity\EntityHandlerInterface;
 use Drupal\Core\Entity\EntityInterface;
+use Drupal\Core\Entity\EntityTypeInterface;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Session\AccountInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Defines the access control handler for the project track tool entity type.
@@ -16,25 +20,37 @@ use Drupal\Core\Session\AccountInterface;
  *
  * @see https://www.drupal.org/project/coder/issues/3185082
  */
-final class ProjectTrackToolAccessControlHandler extends EntityAccessControlHandler {
+final class ProjectTrackToolAccessControlHandler extends EntityAccessControlHandler implements EntityHandlerInterface {
+
+  public function __construct(
+    private readonly EntityTypeManagerInterface $entityTypeManager,
+    EntityTypeInterface $entity_type,
+  ) {
+    parent::__construct($entity_type);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function createInstance(ContainerInterface $container, EntityTypeInterface $entity_type) {
+    return new static(
+      $container->get('entity_type.manager'),
+      $entity_type,
+    );
+  }
 
   /**
    * {@inheritdoc}
    */
   protected function checkAccess(EntityInterface $entity, $operation, AccountInterface $account): AccessResult {
-    if ($account->hasPermission($this->entityType->getAdminPermission())) {
-      return AccessResult::allowed()->cachePerPermissions();
+    if (!($entity instanceof ProjectTrackToolInterface)) {
+      return AccessResult::neutral();
     }
 
-    return match($operation) {
-      'view' => AccessResult::allowedIfHasPermission($account, 'view project_track'),
-      'update' => AccessResult::allowedIfHasPermission($account, 'edit project_track'),
-      'delete' => AccessResult::allowedIfHasPermission($account, 'delete project_track'),
-      'delete revision' => AccessResult::allowedIfHasPermission($account, 'delete project_track revision'),
-      'view all revisions', 'view revision' => AccessResult::allowedIfHasPermissions($account, ['view project_track revision', 'view project_track']),
-      'revert' => AccessResult::allowedIfHasPermissions($account, ['revert project_track revision', 'edit project_track']),
-      default => AccessResult::neutral(),
-    };
+    $parentAccessControlHandler = $this->entityTypeManager->getAccessControlHandler('project_track');
+    $parent = $entity->getProjectTrack();
+
+    return $parentAccessControlHandler->checkAccess($parent, $operation, $account);
   }
 
   /**